fix(dev_react) project structure refactored
